starting equipment feature menu finished

+++ Atelier/src/com/tempera/atelier/dnd/types/equipment/AdventuringGear.java
@@ -1,12 +1,12 @@
 package com.tempera.atelier.dnd.types.equipment;
 
-import com.tempera.atelier.dnd.types.enums.EquipmentType;
 import com.tempera.atelier.dnd.types.enums.Currency.Price;
+import com.tempera.atelier.dnd.types.enums.EquipmentType;
 
 public class AdventuringGear extends BaseItem {
 
 	public AdventuringGear(String name, String description, String source,
-		Price price, int weight) {
+		Price price, double weight) {
 		super(name, description, source, price, weight);
 		addCategories(EquipmentType.ADVENTURING_GEAR);
 	}
